code updates around getting the correct filter component
